Fixed compile of app and server
App wouldn't compile, fixed that because it was still referencing a css page.

For the server, the template in the docs used a middleware to attach the url
Param to the context. I had trouble fetching that and type-casting it, so I
removed the middleware and retreived the param in the main handler.